requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.5
django-admin-inline-paginator==0.1.1
django-debug-toolbar==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytest-env==0.6.2
python-monkey-business==1.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.0
